This script has a function that calculates the smallest prime factor of an integer provided in the command line arguments.
The docstring provides an explanation of how the function works as well as some examples.
"If __name__ == '__main__':" line adds a condition to the script that all codes below it will only run if the file is executed as a script.
Therefore, with if __name__ == '__main__': I can import only the function that is on the top of the script